refactor: S3 test mocking (Bean을 인터페이스로 분리)
