+++ src/test/java/com/devnity/devnity/domain/media/controller/MediaControllerTest.java
@@ -12,6 +12,7 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+import com.devnity.devnity.test.config.MockAwsS3UploaderConfig;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.FileInputStream;
 import org.junit.jupiter.api.Test;
@@ -19,6 +20,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.annotation.Import;
 import org.springframework.core.io.ResourceLoader;
 import org.springframework.http.MediaType;
 import org.springframework.mock.web.MockMultipartFile;
@@ -27,10 +29,13 @@ import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 
+//@ExtendWith(SpringExtension.class)
+@Import(MockAwsS3UploaderConfig.class)
 @WithMockUser(roles = "USER")
 @AutoConfigureRestDocs
 @AutoConfigureMockMvc
 @SpringBootTest
+//@EnableAutoConfiguration(exclude = )
 class MediaControllerTest {
 
   @Autowired
